requirements.txt
pkg-resources==0.0.0
praw==7.1.0
prawcore==1.5.0
python-decouple==3.3
update-checker==0.18.0
websocket-client==0.57.0
